+++ library/docs/changelog.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.1] - 2020-11-12
+
 ### Added
 - Builder pattern for `Laboratory` construction.
 - `DefaultOptionFactory` that can substitute default options for feature flags read by `Laboratory`.
@@ -174,7 +176,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[Unreleased]: https://github.com/MiSikora/Laboratory/compare/0.9.0...HEAD
+[Unreleased]: https://github.com/MiSikora/Laboratory/compare/0.9.1...HEAD
+[0.9.1]: https://github.com/MiSikora/Laboratory/releases/tag/0.9.1
 [0.9.0]: https://github.com/MiSikora/Laboratory/releases/tag/0.9.0
 [0.8.0]: https://github.com/MiSikora/Laboratory/releases/tag/0.8.0
 [0.7.0]: https://github.com/MiSikora/Laboratory/releases/tag/0.7.0